# Conflicts:
#	References/Windows/etc/servers.json

+++ service/servers_updater.go
@@ -1,3 +1,25 @@
+//
+//  Daemon for IVPN Client Desktop
+//  https://github.com/ivpn/desktop-app-daemon
+//
+//  Created by Stelnykovych Alexandr.
+//  Copyright (c) 2020 Privatus Limited.
+//
+//  This file is part of the Daemon for IVPN Client Desktop.
+//
+//  The Daemon for IVPN Client Desktop is free software: you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License as published by the Free
+//  Software Foundation, either version 3 of the License, or (at your option) any later version.
+//
+//  The Daemon for IVPN Client Desktop is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+//  or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+//  details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with the Daemon for IVPN Client Desktop. If not, see <https://www.gnu.org/licenses/>.
+//
+
 package service
 
 import (
@@ -128,7 +150,7 @@ func readServersFromCache() (svrs *types.ServersInfoResponse, apiIPs []string, e
 
 	// check servers.json file has correct access rights (can we use it's data?)
 	mode := stat.Mode()
-	if mode != platform.DefaultFilePermissionForConfig {
+	if platform.DefaultFilePermissionForConfig != 0 && mode != platform.DefaultFilePermissionForConfig {
 		os.Remove(platform.ServersFile())
 		// we can not use servers info from this file
 		// but we can try to get IP addresses of alternate IP's
@@ -153,5 +175,8 @@ func writeServersToCache(servers *types.ServersInfoResponse) error {
 		return errors.New("failed to serialize servers")
 	}
 
-	return ioutil.WriteFile(platform.ServersFile(), data, platform.DefaultFilePermissionForConfig) // only owner (root) can read/write file
+	if platform.DefaultFilePermissionForConfig == 0 {
+		return ioutil.WriteFile(platform.ServersFile(), data, 0600)
+	}
+	return ioutil.WriteFile(platform.ServersFile(), data, platform.DefaultFilePermissionForConfig)
 }
